requirements.txt
aiohttp==3.9.3 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
altair==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
dataclasses-json==0.6.4 ; python_version >= "3.11" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
gitdb==4.0.11 ; python_version >= "3.11" and python_version < "4.0"
gitpython==3.1.41 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.26.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.11" and python_version < "4.0"
jsonpointer==2.4 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.11" and python_version < "4.0"
langchain-community==0.0.19 ; python_version >= "3.11" and python_version < "4.0"
langchain-core==0.1.22 ; python_version >= "3.11" and python_version < "4.0"
langchain-openai==0.0.5 ; python_version >= "3.11" and python_version < "4.0"
langchain==0.1.6 ; python_version >= "3.11" and python_version < "4.0"
langchainhub==0.1.14 ; python_version >= "3.11" and python_version < "4.0"
langsmith==0.0.87 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4.0"
marshmallow==3.20.2 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
neo4j==5.17.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
openai==1.12.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.11" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==15.0.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.16.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
pydeck==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.33.0 ; python_version >= "3.11" and python_version < "4.0"
regex==2023.12.25 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
retry==0.9.2 ; python_version >= "3.11" and python_version < "4.0"
rich==13.7.0 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.17.1 ; python_version >= "3.11" and python_version < "4.0"
segment-analytics-python==2.3.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
smmap==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.25 ; python_version >= "3.11" and python_version < "4.0"
streamlit-chat-media==0.0.4 ; python_version >= "3.11" and python_version < "4.0"
streamlit-chat==0.1.1 ; python_version >= "3.11" and python_version < "4.0"
streamlit-feedback==0.1.3 ; python_version >= "3.11" and python_version < "4.0"
streamlit==1.27.2 ; python_version >= "3.11" and python_version < "4.0"
tenacity==8.2.3 ; python_version >= "3.11" and python_version < "4.0"
tiktoken==0.5.2 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
toolz==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.4 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.11" and python_version < "4.0"
types-requests==2.31.0.20240125 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.4 ; python_version >= "3.11" and python_version < "4.0"
tzlocal==5.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
validators==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
watchdog==4.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_system != "Darwin"
yarl==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.11" and python_version < "4.0"
